project/requirements.txt
flask==2.3.3
PyPDF2==3.0.1
openai==1.3.7
python-dotenv==1.0.0
tiktoken==0.5.1
faiss-cpu==1.8.0
transformers==4.35.0
torch==2.2.0
sentence-transformers
pandas
numpy
PyMuPDF
